added frontend and backend folders
in order to prepare for a basic front-end version of the app, I put all back-end code into the backend/ folder, and will put the flutter front-end in the frontend/ folder. In the future it may be warranted to make a second repository for the front-end, but to simplify development, I will keep them separated but within the same repository for now.